+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/oauth-mock-server
 
+## 2.5.30
+
+### Patch Changes
+
+- More OAuth providers
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.30
+
+## 2.5.29
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.5.29
+
 ## 2.5.28
 
 ### Patch Changes