+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackframe/stack-shared
 
+## 2.5.30
+
+### Patch Changes
+
+- More OAuth providers
+  - @stackframe/stack-sc@2.5.30
+
+## 2.5.29
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.5.29
+
 ## 2.5.28
 
 ### Patch Changes